+++ src/com/org/model/classes/agravos/ContatosExaminadosHanseniasePactuacao.java
@@ -582,7 +582,7 @@ public class ContatosExaminadosHanseniasePactuacao extends Agravo {
                 rowData[7] = String.format("%.1f", percentual);
             }
 
-            rowData[8] = String.valueOf(preencheAno(getDataInicio(), getDataFim()));
+            rowData[8] = String.valueOf(preencheAno(getDataInicio(), getDataFim()).intValue());
             rowData[9] = getDataInicio();
             rowData[10] = getDataFim();
             rowData[11] = "HANSENIASE-SINANNET";
